refactor modules properly
