+++ futurecommander_gui/app_web/common/icon.js
