+++ futurecommander_gui/app_web/index.html
@@ -17,4 +17,4 @@
   ~ along with FutureCommander.  If not, see <https://www.gnu.org/licenses/>.
   -->
 <!DOCTYPE html>
-<link rel="stylesheet" href="view/style.css">
+<link rel="stylesheet" href="common/style.css">
